test: split schema gen benchmarks for recursive models into separate file
